requirements.txt
fastapi==0.115.8
uvicorn==0.15.0
python-multipart==0.0.20
llama-index==0.12.14
llama-index-llms-openai==0.3.14
llama-index-readers-json==0.3.0
llama-index-vector-stores-pinecone==0.4.4
pinecone-client==5.0.1
python-docx==1.1.2
docx2txt==0.8
openai==1.60.2
python-dotenv==1.0.1
PyPDF2==3.0.1